Vulkan: Add Memory Usage Tracking

+++ OpenVoxel/src/main/java/net/openvoxel/client/renderer/vk/VkWorldRenderer.java
@@ -61,7 +61,7 @@ public class VkWorldRenderer implements WorldRenderer {
 	/**
 	 * Must be called in vulkan exclusive context [after wait idle]
 	 */
-	public void shrinkMemory(int targetAmount) {
+	public void shrinkMemory(long targetAmount) {
 		int limit = memoryManager.reclaim_memory(targetAmount);
 		//TODO: mark those that no-longer have state//
 	}
